Make routes modular... Still not happy with how auth redirection is being handled...
