+++ views/profile.pug
@@ -4,4 +4,4 @@ block content
     h1 Welcome!
     .user #{identity + ' ' + score}
         p
-            a(href="/logout") logout
+            a(href="/app/logout") logout
